% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invcov2pcor.R
\name{invcov2pcor}
\alias{invcov2pcor}
\title{Transforms an inverse covariance matrix into a partial correlation matrix}
\usage{
invcov2pcor(mat)
}
\arguments{
\item{mat}{the given inverse covariance matrix.}
}
\value{
The corresponding partial correlation matrix.
}
\description{
Transforms the positive definite \eqn{p \times p} input precision matrix into the
corresponding \eqn{p \times p} partial correlation matrix.
}
\references{
Whittaker, J. (1990). *Graphical models in applied multivariate statistics*. Wiley.
}
\keyword{algebra}
\keyword{multivariate}
