% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isComplete.R
\name{isComplete}
\alias{isComplete}
\title{Test if a given (sub)graph is complete}
\usage{
isComplete(G, set)
}
\arguments{
\item{G}{an \code{\link[igraph:igraph-package]{igraph}} object.}

\item{set}{vector of the subgraph's nodes.}
}
\value{
Boolean: TRUE if the (sub)graph is complete, FALSE if not.
}
\description{
Determines if a given graph is complete or not.
}
\keyword{graphs}
