\name{isDecomposable}
\alias{isDecomposable}
\title{Tests if a given graph is decomposable}
\description{Determines if a given graph is decomposable, or chordal, or not.}
\usage{isDecomposable(G)}
\arguments{
  \item{G}{an \code{\link[igraph0:igraph-package]{igraph0}} object.}
}
\value{Boolean : TRUE if the graph is decompsable, FALSE if not.}
\keyword{ graphs }
