\name{invcov2pcor}
\alias{invcov2pcor}
\title{Transforms an inverse covariance matrix into a partial correlation matrix}
\description{Transforms the positive definite p x p input precision matrix into the corresponding p x p partial correlation matrix.}
\usage{invcov2pcor(mat)}
\arguments{
  \item{mat}{the given inverse covariance matrix.}
}
\value{The corresponding partial correlation matrix.}
\references{Whittaker, J. (1990). Graphical models in applied multivariate statistics. Wiley.}
\keyword{ algebra }
\keyword{ multivariate }
