\name{isComplete}
\alias{isComplete}
\title{Test if a given (sub)graph is complete}
\description{Determines if a given graph is complete or not.}
\usage{isComplete(G,set)}
\arguments{
  \item{G}{an \code{\link[igraph:igraph-package]{igraph}} object.}
  \item{set}{vector of the subgraph's nodes.}
}
\value{Boolean : TRUE if the (sub)graph is complete, FALSE if not.}
\keyword{ graphs }
